fix: move RepositoryLimits to own sub-package for DB isolation
The ProvisioningMaxRepositories=10 setting on the shared server caused
RepositoryValidation subtests to hit the limit (they create 10+ repos
across subtests). It also caused RepositoryLimits itself to see stale
quota counts from the shared DB.

Moving the test to its own sub-package gives it a separate TestMain,
test database, and Grafana server instance — the isolation it needs.

Made-with: Cursor
