+++ pkg/tests/apis/provisioning/repository/repository_test.go
@@ -1120,131 +1120,6 @@ func TestIntegrationProvisioning_WebhookRejectedForUnhealthyRepository(t *testin
 	require.Error(t, result.Error(), "webhook request should be rejected for unhealthy repository")
 	require.Equal(t, http.StatusFailedDependency, statusCode, "should return 424 Failed Dependency for unhealthy repository")
 }
-func TestIntegrationProvisioning_RepositoryLimits(t *testing.T) {
-	helper := sharedHelper(t)
-	ctx := context.Background()
-
-	originalName := "original-repo"
-	// Create instance sync repository first
-	originalRepo := common.TestRepo{
-		Name:               originalName,
-		Target:             "instance",
-		Copies:             map[string]string{}, // No files needed for this test
-		ExpectedDashboards: 0,
-		ExpectedFolders:    0,
-	}
-	helper.CreateRepo(t, originalRepo)
-
-	t.Run("folder sync is rejected when instance sync exists", func(t *testing.T) {
-		folderRepo := helper.RenderObject(t, "../testdata/local-write.json.tmpl", map[string]any{
-			"Name":        "folder-blocked-by-instance",
-			"SyncEnabled": true,
-			"SyncTarget":  "folder",
-		})
-
-		_, err := helper.Repositories.Resource.Create(ctx, folderRepo, metav1.CreateOptions{FieldValidation: "Strict"})
-		require.Error(t, err, "folder sync repository should be rejected when instance sync exists")
-
-		// Verify the error message mentions the existing instance repository
-		statusError := helper.RequireApiErrorStatus(err, metav1.StatusReasonInvalid, http.StatusUnprocessableEntity)
-		require.Contains(t, statusError.Message, "Cannot create folder repository when instance repository exists: "+originalName)
-	})
-
-	t.Run("change between folder and instance sync for the same repository if no previous sync happened", func(t *testing.T) {
-		repo, err := helper.Repositories.Resource.Get(ctx, originalName, metav1.GetOptions{})
-		require.NoError(t, err, "failed to get repository")
-		err = unstructured.SetNestedField(repo.Object, "folder", "spec", "sync", "target")
-		require.NoError(t, err, "failed to set syncTarget to folder")
-		_, err = helper.Repositories.Resource.Update(ctx, repo, metav1.UpdateOptions{FieldValidation: "Strict"})
-		require.NoError(t, err, "failed to update repository to folder sync")
-
-		// Verify that the repository is now a folder sync
-		// We verify with the listing APIs because it may take some time for the update to propagate
-		require.Eventually(t, func() bool {
-			repos, err := helper.Repositories.Resource.List(ctx, metav1.ListOptions{})
-			if err != nil {
-				return false
-			}
-
-			for _, repo := range repos.Items {
-				if repo.GetName() == originalName {
-					syncTarget, found, err := unstructured.NestedString(repo.Object, "spec", "sync", "target")
-					if err != nil || !found {
-						return false
-					}
-
-					return syncTarget == "folder"
-				}
-			}
-
-			return false
-		}, time.Second*10, time.Millisecond*100, "failed to verify that sync target is folder")
-	})
-
-	t.Run("instance sync rejected when any other repository exists", func(t *testing.T) {
-		instanceRepo := helper.RenderObject(t, "../testdata/local-write.json.tmpl", map[string]any{
-			"Name":        "instance-repo-blocked",
-			"SyncEnabled": true,
-			"SyncTarget":  "instance",
-		})
-
-		_, err := helper.Repositories.Resource.Create(ctx, instanceRepo, metav1.CreateOptions{FieldValidation: "Strict"})
-		require.Error(t, err, "instance sync repository should be rejected when any other repository exists")
-
-		statusError := helper.RequireApiErrorStatus(err, metav1.StatusReasonInvalid, http.StatusUnprocessableEntity)
-		require.Contains(t, statusError.Message, "Instance repository can only be created when no other repositories exist. Found: "+originalName)
-	})
-
-	t.Run("repository limit validation of 10 for folder syncs repositories", func(t *testing.T) {
-		// Ensure the original repo is folder sync before testing limits
-		// (it was changed to folder sync in a previous subtest)
-		require.Eventually(t, func() bool {
-			repo, err := helper.Repositories.Resource.Get(ctx, originalName, metav1.GetOptions{})
-			if err != nil {
-				return false
-			}
-			syncTarget, found, err := unstructured.NestedString(repo.Object, "spec", "sync", "target")
-			if err != nil || !found {
-				return false
-			}
-			return syncTarget == "folder"
-		}, time.Second*10, time.Millisecond*100, "original repo should be folder sync")
-
-		// Count existing repos (should be 1: original-repo)
-		existingRepos, err := helper.Repositories.Resource.List(ctx, metav1.ListOptions{})
-		require.NoError(t, err, "failed to list repositories")
-		existingCount := len(existingRepos.Items)
-		require.Equal(t, 1, existingCount, "should have 1 existing repository (original-repo)")
-
-		// Create repos to reach the limit of 10 (we already have 1, so create 9 more)
-		for i := 2; i <= 10; i++ {
-			repoName := fmt.Sprintf("limit-test-repo-%d", i)
-			limitTestRepo := common.TestRepo{
-				Name:               repoName,
-				Target:             "folder",
-				Copies:             map[string]string{}, // No files needed for this test
-				ExpectedDashboards: 0,
-				ExpectedFolders:    i, // Each repository creates a folder, so total = i
-			}
-			helper.CreateRepo(t, limitTestRepo)
-		}
-
-		// Try to create the 11th repository - should fail due to limit
-		eleventhRepoName := "limit-test-repo-11"
-		eleventhRepo := helper.RenderObject(t, "../testdata/local-write.json.tmpl", map[string]any{
-			"Name":        eleventhRepoName,
-			"SyncEnabled": true,
-			"SyncTarget":  "folder",
-		})
-
-		_, createErr := helper.Repositories.Resource.Create(ctx, eleventhRepo, metav1.CreateOptions{FieldValidation: "Strict"})
-		require.Error(t, createErr, "11th repository should be rejected due to limit")
-
-		statusError := helper.RequireApiErrorStatus(createErr, metav1.StatusReasonInvalid, http.StatusUnprocessableEntity)
-		require.Contains(t, statusError.Message, "Maximum number of 10 repositories reached")
-	})
-}
-
 func TestIntegrationProvisioning_RunLocalRepository(t *testing.T) {
 	helper := sharedHelper(t)
 	ctx := context.Background()
